Automatically encode input in log functions
